\name{gaussLoad}
\alias{gaussLoad}
\title{Load a GAUSS object.}
\usage{
  gaussLoad(filename)
}
\arguments{
  \item{filename}{Load gauss list from this file}
}
\value{
  list \code{gauss}
}
\description{
  Load a GAUSS object.
}
\references{
  \code{\link{gaussSave}}
}
\keyword{internal}

