% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcov.R
\name{lcovCreate}
\alias{lcovCreate}
\title{Create a new covariance object.}
\usage{
lcovCreate(dim)
}
\arguments{
\item{dim}{dimension of the covariance object list to be created}
}
\value{
list containing the COVMTX, avg and tlen
}
\description{
Create a new covariance object.
}
\seealso{
\code{\link{lcovUpdate}}
}
\keyword{internal}
