% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\docType{package}
\name{rSFA-package}
\alias{rSFA-package}
\alias{rSFA}
\title{Slow Feature Analysis}
\description{
Slow Feature Analysis
}
\details{
\tabular{ll}{
Package: \tab rSFA\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 29.03.2022\cr
Maintainer: \tab Martin Zaefferer \email{martin.zaefferer@gmx.de}\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Slow Feature Analysis (SFA), ported to R based on the matlab implementations SFA toolkit 1.0 by Pietro Berkes and SFA
   toolkit 2.8 by Wolfgang Konen.
}
\author{
Wolfgang Konen \email{wolfgang.konen@fh-koeln.de}, Martin Zaefferer, Patrick Koch; 
		Bug hunting and testing by Ayodele Fasika, Ashwin Kumar, Prawyn Jebakumar
}
\keyword{analysis}
\keyword{classification}
\keyword{feature}
\keyword{slow}
\keyword{timeseries}
