\name{sfaBSh}
\alias{sfaBSh}
\title{Backslash operator}
\usage{
  sfaBSh(X, Y)
}
\arguments{
  \item{X}{X matrix}

  \item{Y}{Y vector}
}
\value{
  Returns coefficients
}
\description{
  Reproduce what matlabs backslash operator can do, using
  qr() and qr.coef().
}
\keyword{internal}

