% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sfaGetIntRange}
\alias{sfaGetIntRange}
\title{Helper Function of SFA.}
\usage{
sfaGetIntRange(range)
}
\arguments{
\item{range}{Range}
}
\value{
numeric vector \code{int}
}
\description{
Helper Function of SFA.
}
\keyword{internal}

