\name{sfaLoad}
\alias{sfaLoad}
\title{Load a SFA object.}
\usage{
  sfaLoad(filename)
}
\arguments{
  \item{filename}{Load sfa list from this file name}
}
\value{
  list \code{sfaList}
}
\description{
  Load a SFA object.
}
\references{
  \code{\link{sfaSave}}
}
\keyword{internal}

