\name{sfaPreproc}
\alias{sfaPreproc}
\title{Preprocessing for sfa classification}
\usage{
  sfaPreproc(sfaList, x, opts)
}
\arguments{
  \item{sfaList}{A list that contains all information about
  the handled sfa-structure}

  \item{x}{Input data, each column a different variable}

  \item{opts}{list}
}
\value{
  preprocessed data
}
\description{
  Preprocessing for sfa classification
}
\references{
  \code{\link{sfaClassify}}
}

