\name{xpDim}
\alias{xpDim}
\title{Degree 2 Dimension Calculation}
\usage{
  xpDim(n)
}
\arguments{
  \item{n}{Dimension of input vector}
}
\value{
  Dimension of expanded vector
}
\description{
  Compute the dimension of a vector expanded in the space
  of polynomials of 2nd degree.
}
\references{
  \code{\link{sfa2}} \code{\link{sfaExpand}}
}

