\name{customRep}
\alias{customRep}
\title{Custom Repeater Function}
\usage{
  customRep(a, n)
}
\arguments{
  \item{a}{to be repeated}

  \item{n}{repeat by}
}
\value{
  Returns the repeated dataset
}
\description{
  Faster than customRepmat in matlab package, in certain
  cases.
}
\keyword{internal}

