% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.R
\name{gaussCreate}
\alias{gaussCreate}
\title{Create an Gaussian classifier object}
\usage{
gaussCreate(nclass, dimY)
}
\arguments{
\item{nclass}{number of classes}

\item{dimY}{dimension}
}
\value{
list of defaults for gauss classifier
}
\description{
Create an Gaussian classifier object
}
\seealso{
\code{\link{gaussClassifier}}
}
