% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gaussLoad}
\alias{gaussLoad}
\title{Load a GAUSS object.}
\usage{
gaussLoad(filename)
}
\arguments{
\item{filename}{Load list \code{gauss} from this file}
}
\value{
list \code{gauss}
}
\description{
Load a GAUSS object.
}
\references{
\code{\link{gaussSave}}
}
\keyword{internal}

