% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfaFileIO.R
\name{gaussLoad}
\alias{gaussLoad}
\title{Load a GAUSS object.}
\usage{
gaussLoad(filename)
}
\arguments{
\item{filename}{Load list \code{gauss} from this file}
}
\value{
list \code{gauss}
}
\description{
Load a GAUSS object.
}
\references{
\code{\link{gaussSave}}
}
\keyword{internal}
