\name{gaussSave}
\alias{gaussSave}
\title{Save a GAUSS object.}
\usage{
  gaussSave(gauss, filename)
}
\arguments{
  \item{gauss}{A list that contains all information about
  the handled gauss-structure}

  \item{filename}{Save gauss list to this file}
}
\description{
  Save a GAUSS object.
}
\references{
  \code{\link{gaussLoad}}
}
\keyword{internal}

