% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gaussSave}
\alias{gaussSave}
\title{Save a GAUSS object.}
\usage{
gaussSave(gauss, filename)
}
\arguments{
\item{gauss}{A list that contains all information about the handled gauss-structure}

\item{filename}{Save list \code{gauss} to this file}
}
\description{
Save a GAUSS object.
}
\references{
\code{\link{gaussLoad}}
}
\keyword{internal}

