% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfaExpand.R
\name{nlDim}
\alias{nlDim}
\title{Custom Nonlinear Dimension Calculation}
\usage{
nlDim(n)
}
\arguments{
\item{n}{Dimension of input vector}
}
\value{
Dimension of expanded vector
}
\description{
Compute the dimension, used for the customexpansion demo.
}
\seealso{
\code{\link{sfa2}} \code{\link{nlExpand}}
}
\keyword{internal}
