\docType{package}
\name{rSFA-package}
\alias{rSFA}
\alias{rSFA-package}
\title{Slow Feature Analysis in R}
\description{
  Slow Feature Analysis in R
}
\details{
  \tabular{ll}{ Package: \tab rSFA\cr Type: \tab Package\cr
  Version: \tab 1.02\cr Date: \tab 20.11.2012\cr
  Maintainer: \tab Martin Zaefferer
  \email{martin.zaefferer@gmx.de}\cr License: \tab GPL (>=
  2)\cr LazyLoad: \tab yes\cr }

  Slow Feature Analysis in R, ported to R based on the
  matlab versions SFA toolkit 1.0 by Pietro Berkes and SFA
  toolkit 2.8 by Wolfgang Konen.
}
\author{
  Wolfgang Konen \email{wolfgang.konen@fh-koeln.de}, Martin
  Zaefferer, Patrick Koch; Bug hunting and testing by
  Ayodele Fasika, Ashwin Kumar, Prawyn Jebakumar
}
\references{
  \url{http://gociop.de/research-projects/sfa/}
}
\keyword{analysis}
\keyword{classification}
\keyword{feature}
\keyword{slow}
\keyword{timeseries}

