% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sfa1Create}
\alias{sfa1Create}
\title{Create structured list for linear SFA}
\usage{
sfa1Create(sfaRange, axType = "ORD1", regCt = 0)
}
\arguments{
\item{sfaRange}{number of slowly-varying functions to be kept}

\item{axType}{is the type of derivative approximation to be used, see \code{\link{sfaTimediff}}}

\item{regCt}{regularization constant, currently not used}
}
\value{
list \code{sfaList} contains all arguments passed into sfa1create plus
   \item{deg}{ 2}
This list will be expanded by other SFA functions with further SFa results
}
\description{
Create structured list for linear SFA
}
\seealso{
\code{\link{sfa1}} \code{\link{sfaStep}} \code{\link{sfa2Create}}
}

