% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfaClassify.R
\name{sfaClassPredict}
\alias{sfaClassPredict}
\title{Predict Class for SFA classification}
\usage{
sfaClassPredict(xtst, realcTst, opts)
}
\arguments{
\item{xtst}{NTST x IDIM, test input data}

\item{realcTst}{1 x NTST, test class labels}

\item{opts}{list with several parameter settings: \describe{
                      \item{gaussdim}{}
                   \item{    ... }{}
                      \item{ *Filename}{ [* = s,g,x] from where to load the models (see \code{\link{sfaClassify}})   }
}}
}
\value{
list \code{res} containing \cr
   \item{res$errtst}{ 1 x 2 matrix: error rate with / w/o SFA on test set }
   \item{res$ytst}{ output from SFA when applied to test data  }
   \item{res$predT}{ predictions with SFA + GaussClassifier on test set }
   \item{res$predX}{ predictions w/o SFA (only GaussClassifier) on test set (only if opts.xFilename exists) }
}
\description{
Use a SFA classification model (stored in opts$*Filename), predict & evaluate on new data (xtst,realc_tst).\cr
Author of orig. matlab version: Wolfgang Konen, Jan 2011-Mar 2011.\cr
See also [Berkes05] Pietro Berkes: Pattern recognition with Slow Feature Analysis. 
Cognitive Sciences EPrint Archive (CogPrint) 4104, http://cogprints.org/4104/ (2005)
}
\seealso{
\code{\link{sfaClassify}} \code{\link{sfaExecute}}
}
