% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfaHelperFunctions.R
\name{sfaGetIntRange}
\alias{sfaGetIntRange}
\title{Helper Function of SFA.}
\usage{
sfaGetIntRange(range)
}
\arguments{
\item{range}{Range}
}
\value{
numeric vector \code{int}
}
\description{
Helper Function of SFA.
}
\keyword{internal}
