% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfaFileIO.R
\name{sfaLoad}
\alias{sfaLoad}
\title{Load a SFA object.}
\usage{
sfaLoad(filename)
}
\arguments{
\item{filename}{Load list \code{sfaList} from this file name}
}
\value{
list \code{sfaList}
}
\description{
Load a SFA object.
}
\references{
\code{\link{sfaSave}}
}
\keyword{internal}
